use ints not shorts
git-svn-id: https://nocs.stanford.edu/cgi-bin/svn.cgi/booksim2.0/trunk@2777 ce7aba0d-9f81-40a6-a495-cecae84b0515

+++ src/flit.hpp
@@ -72,7 +72,7 @@ public:
 
   int  hops;
   bool watch;
-  short subnetwork;
+  int  subnetwork;
 
   //for credit tracking, last router visited
   mutable int from_router;
