+++ src/networks/flatfly_onchip.hpp
@@ -69,7 +69,7 @@ public:
   double Capacity( ) const;
   void InsertRandomFaults( const Configuration &config );
 
-  static short half_vcs;
+  static int half_vcs;
 };
 
 void xyyx_flatfly( const Router *r, const Flit *f, int in_channel,
